requirements.txt
streamlit
google-genai
edge-tts
google-cloud-texttospeech
requests
google-auth
elevenlabs>=1.0.0
pypdf

